# api/openapi.policy.json
{
  "openapi" : "3.1.0",
  "info" : {
    "title" : "OpenAPI definition",
    "version" : "v0"
  },
  "servers" : [ {
    "url" : "http://localhost",
    "description" : "Generated server url"
  } ],
  "paths" : {
...
    "/api/v1/policies" : {
      "get" : {
        "operationId" : "listPolicies",
        "parameters" : [ {
          "in" : "query",
          "name" : "q",
          "required" : false,
          "schema" : {
            "type" : "string",
            "default" : "",
            "maxLength" : 80,
            "minLength" : 0
          }
        }, {
          "in" : "query",
          "name" : "page",
          "required" : false,
          "schema" : {
            "type" : "integer",
            "format" : "int32",
            "default" : 1,
            "maximum" : 1000,
            "minimum" : 1
          }
        }, {
          "in" : "query",
          "name" : "pageSize",
          "required" : false,
          "schema" : {
            "type" : "integer",
            "format" : "int32",
            "default" : 20,
            "maximum" : 50,
            "minimum" : 1
          }
        }, {
          "in" : "query",
          "name" : "questionsOnly",
          "required" : false,
          "schema" : {
            "type" : "boolean",
            "default" : false
          }
        }, {
          "in" : "query",
          "name" : "recruitmentStatus",
          "required" : false,
          "schema" : {
            "type" : "string",
            "enum" : [ "BEFORE_OPENING", "OPEN", "CLOSED", "ROLLING", "UNTIL_EXHAUSTED", "UNKNOWN" ]
          }
        }, {
          "in" : "query",
          "name" : "category",
          "required" : false,
          "schema" : {
            "type" : "array",
            "items" : {
              "type" : "string",
              "enum" : [ "JOB", "HOUSING", "EDUCATION", "FINANCE", "PARTICIPATION" ]
            }
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/PolicyListResponse"
                }
              }
            },
            "description" : "OK"
          },
          "400" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/PolicyApiError"
                }
              }
            },
            "description" : "Bad Request"
          },
          "503" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/PolicyApiError"
                }
              }
            },
            "description" : "정책 저장소 조회 실패"
          }
        },
        "summary" : "정책 검색·분야(여러 개는 하나라도 해당)·질문 제공 여부·접수 상태 필터. 접수 중인 정책을 먼저, 마감 임박순으로 정렬",
        "tags" : [ "policy-catalog-controller" ]
      }
    },
    "/api/v1/policies/category-counts" : {
      "get" : {
        "operationId" : "countPoliciesByCategory",
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/PolicyCategoryCounts"
                }
              }
            },
            "description" : "OK"
          },
          "503" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/PolicyApiError"
                }
              }
            },
            "description" : "정책 저장소 조회 실패"
          }
        },
        "summary" : "분야별 공개 정책 수. 목록의 분야 필터와 같은 기준",
        "tags" : [ "policy-catalog-controller" ]
      }
    },
    "/api/v1/policies/checks" : {
      "post" : {
        "operationId" : "checkPolicyConditions",
        "parameters" : [ {
          "in" : "query",
          "name" : "page",
          "required" : false,
          "schema" : {
            "type" : "integer",
            "format" : "int32",
            "default" : 1,
            "maximum" : 1000,
            "minimum" : 1
          }
        }, {
          "in" : "query",
          "name" : "q",
          "required" : false,
          "schema" : {
            "type" : "string",
            "default" : "",
            "maxLength" : 80,
            "minLength" : 0
          }
        }, {
          "in" : "query",
          "name" : "sort",
          "required" : false,
          "schema" : {
            "type" : "string",
            "default" : "AGE_MATCH",
            "enum" : [ "AGE_MATCH", "RECENT" ]
          }
        }, {
          "in" : "query",
          "name" : "recruitmentStatus",
          "required" : false,
          "schema" : {
            "type" : "string",
            "enum" : [ "BEFORE_OPENING", "OPEN", "CLOSED", "ROLLING", "UNTIL_EXHAUSTED", "UNKNOWN" ]
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/BasicConditions"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/PolicyCheckResponse"
                }
              }
            },
            "description" : "OK"
          },
          "400" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/PolicyApiError"
                }
              }
            },
            "description" : "Bad Request"
          },
          "503" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/PolicyApiError"
                }
              }
            },
            "description" : "정책 저장소 조회 실패"
          }
        },
        "summary" : "기본 조건으로 연령 비교·정책 검색·접수 상태 필터·정렬. 조건은 저장하지 않음",
        "tags" : [ "policy-catalog-controller" ]
      }
    },
    "/api/v1/policies/{number}/evaluation" : {
      "post" : {
        "operationId" : "evaluatePolicyAnswers",
        "parameters" : [ {
          "in" : "path",
          "name" : "number",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/PolicyEvaluationRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/PolicyEvaluation"
                }
              }
            },
            "description" : "OK"
          },
          "400" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/PolicyApiError"
                }
              }
            },
            "description" : "질문·답변 형식 오류"
          },
          "404" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/PolicyApiError"
                }
              }
            },
            "description" : "정책 없음"
          },
          "409" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/PolicyApiError"
                }
              }
            },
            "description" : "정책 개정 또는 질문 변경"
          },
          "503" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/PolicyApiError"
                }
              }
            },
            "description" : "정책 저장소 조회 실패"
          }
        },
        "summary" : "추가 답변으로 확인한 공통요건 비교. 답변은 저장하지 않음",
        "tags" : [ "policy-question-controller" ]
      }
    },
    "/api/v1/policies/{number}/question-prefill" : {
      "post" : {
        "operationId" : "prefillPolicyAnswers",
        "parameters" : [ {
          "in" : "path",
          "name" : "number",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/PolicyPrefillRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/PolicyPrefill"
                }
              }
            },
            "description" : "OK"
          },
          "400" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/PolicyApiError"
                }
              }
            },
            "description" : "생년월일 형식 오류"
          },
          "404" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/PolicyApiError"
                }
              }
            },
            "description" : "정책 없음"
          },
          "409" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/PolicyApiError"
                }
              }
            },
            "description" : "정책 개정 또는 질문 변경"
          },
          "503" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/PolicyApiError"
                }
              }
            },
            "description" : "정책 저장소 조회 실패"
          }
        },
        "summary" : "확인한 생년월일을 현재 공고의 출생일 답변으로 변환. 저장하지 않음",
        "tags" : [ "policy-question-controller" ]
      }
    },
    "/api/v1/policies/{number}/questions" : {
      "get" : {
        "operationId" : "getPolicyQuestions",
        "parameters" : [ {
          "in" : "path",
          "name" : "number",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/PolicyQuestionnaire"
                }
              }
            },
            "description" : "OK"
          },
          "404" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/PolicyApiError"
                }
              }
            },
            "description" : "정책 없음"
          },
          "503" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/PolicyApiError"
                }
              }
            },
            "description" : "정책 저장소 조회 실패"
          }
        },
        "summary" : "현재 정책 개정에서 검토한 추가 질문 조회",
        "tags" : [ "policy-question-controller" ]
      }
    },
    "/api/v1/policies/{policyNumber}" : {
      "get" : {
        "operationId" : "getPolicy",
        "parameters" : [ {
          "in" : "path",
          "name" : "policyNumber",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/PolicyDetailResponse"
                }
              }
            },
            "description" : "OK"
          },
          "404" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/PolicyApiError"
                }
              }
            },
            "description" : "Not Found"
          },
          "503" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/PolicyApiError"
                }
              }
            },
            "description" : "정책 저장소 조회 실패"
          }
        },
        "summary" : "정책 상세와 원문 출처 조회",
        "tags" : [ "policy-catalog-controller" ]
      }
    },
...
  },
  "components" : {
    "schemas" : {
...
      "BasicConditions" : {
        "type" : "object",
        "properties" : {
          "birthDate" : {
            "type" : [ "string", "null" ],
            "format" : "date"
          },
          "district" : {
            "type" : [ "string", "null" ],
            "enum" : [ "강남구", "강동구", "강북구", "강서구", "관악구", "광진구", "구로구", "금천구", "노원구", "도봉구", "동대문구", "동작구", "마포구", "서대문구", "서초구", "성동구", "성북구", "송파구", "양천구", "영등포구", "용산구", "은평구", "종로구", "중구", "중랑구", null ]
          },
          "employmentStatus" : {
            "type" : [ "string", "null" ],
            "enum" : [ "EMPLOYED", "SELF_EMPLOYED", "NOT_EMPLOYED", "FREELANCER", "DAY_WORKER", "ENTREPRENEUR", "SHORT_TERM_WORKER", "FARMER", "OTHER", null ]
          }
        }
      },
...
      "PolicyAnswer" : {
        "type" : "object",
        "properties" : {
          "questionId" : {
            "type" : "string",
            "maxLength" : 60,
            "minLength" : 0
          },
          "value" : {
            "type" : "string",
            "maxLength" : 40,
            "minLength" : 0
          }
        },
        "required" : [ "questionId", "value" ]
      },
      "PolicyAnswerOption" : {
        "type" : "object",
        "properties" : {
          "label" : {
            "type" : "string",
            "minLength" : 1
          },
          "value" : {
            "type" : "string",
            "maxLength" : 40,
            "minLength" : 0
          }
        },
        "required" : [ "label", "value" ]
      },
      "PolicyApiError" : {
        "type" : "object",
        "properties" : {
          "code" : {
            "type" : "string"
          },
          "message" : {
            "type" : "string"
          }
        },
        "required" : [ "code", "message" ]
      },
      "PolicyCategoryCount" : {
        "type" : "object",
        "properties" : {
          "category" : {
            "type" : "string",
            "enum" : [ "JOB", "HOUSING", "EDUCATION", "FINANCE", "PARTICIPATION" ]
          },
          "count" : {
            "type" : "integer",
            "format" : "int64"
          }
        },
        "required" : [ "category", "count" ]
      },
      "PolicyCategoryCounts" : {
        "type" : "object",
        "properties" : {
          "items" : {
            "type" : "array",
            "description" : "다섯 분야를 항상 같은 순서로 제공. 정책이 없으면 0. 복수 분류 정책은 분야마다 세므로 합이 total보다 클 수 있음",
            "items" : {
              "$ref" : "#/components/schemas/PolicyCategoryCount"
            }
          },
          "total" : {
            "type" : "integer",
            "format" : "int64",
            "description" : "전체 공개 정책 수. 다섯 분야에 맞지 않는 정책도 포함"
          }
        },
        "required" : [ "items", "total" ]
      },
      "PolicyCheckItem" : {
        "type" : "object",
        "properties" : {
          "applicationPeriod" : {
            "type" : "string"
          },
          "checks" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/PolicyEvaluatedCheck"
            }
          },
          "collectedAt" : {
            "type" : "string",
            "format" : "date-time"
          },
          "explanation" : {
            "type" : "string"
          },
          "policyNumber" : {
            "type" : "string"
          },
          "questionnaireAvailable" : {
            "type" : "boolean",
            "description" : "현재 원문·적용 시점에 맞는 공통요건 질문 제공 여부. 자격·모집 상태와 무관"
          },
          "recruitment" : {
            "$ref" : "#/components/schemas/PolicyRecruitment"
          },
          "revision" : {
            "type" : "integer",
            "format" : "int64"
          },
          "ruleVersion" : {
            "type" : "string",
            "description" : "적용한 연령 규칙 버전. 검토된 비교 기준이 없으면 빈 문자열"
          },
          "sourceUrl" : {
            "type" : "string"
          },
          "status" : {
            "type" : "string",
            "enum" : [ "ELIGIBLE", "NEEDS_REVIEW", "INELIGIBLE" ]
          },
          "title" : {
            "type" : "string"
          }
        },
        "required" : [ "applicationPeriod", "checks", "collectedAt", "explanation", "policyNumber", "questionnaireAvailable", "recruitment", "revision", "ruleVersion", "sourceUrl", "status", "title" ]
      },
      "PolicyCheckResponse" : {
        "type" : "object",
        "properties" : {
          "evaluatedAt" : {
            "type" : "string",
            "format" : "date-time"
          },
          "hasNext" : {
            "type" : "boolean"
          },
          "items" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/PolicyCheckItem"
            }
          },
          "page" : {
            "type" : "integer",
            "format" : "int32"
          },
          "total" : {
            "type" : "integer",
            "format" : "int64"
          }
        },
        "required" : [ "evaluatedAt", "hasNext", "items", "page", "total" ]
      },
      "PolicyContent" : {
        "type" : "object",
        "properties" : {
          "applicationPeriod" : {
            "type" : "string"
          },
          "category" : {
            "type" : "string"
          },
          "description" : {
            "type" : "string"
          },
          "links" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/PolicyOfficialLink"
            }
          },
          "organization" : {
            "type" : "string"
          },
          "regionCodes" : {
            "type" : "array",
            "items" : {
              "type" : "string"
            }
          },
          "sections" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/PolicyTextSection"
            }
          },
          "sourceModifiedAtText" : {
            "type" : "string",
            "description" : "시간대가 확정되지 않은 원천 수정 일시 원문"
          },
          "title" : {
            "type" : "string"
          }
        },
        "required" : [ "applicationPeriod", "category", "description", "links", "organization", "regionCodes", "sections", "sourceModifiedAtText", "title" ]
      },
...
      "PolicyDetailResponse" : {
        "type" : "object",
        "properties" : {
          "collectedAt" : {
            "type" : "string",
            "format" : "date-time"
          },
          "content" : {
            "$ref" : "#/components/schemas/PolicyContent"
          },
          "policyNumber" : {
            "type" : "string"
          },
          "recruitment" : {
            "$ref" : "#/components/schemas/PolicyRecruitment"
          },
          "revision" : {
            "type" : "integer",
            "format" : "int64",
            "description" : "서비스 내부 적용 개정 번호"
          },
          "sourceConditions" : {
            "type" : "array",
            "description" : "온통청년에 등록된 조건 표기. 자격 판정이 아니며 제한을 표기한 값만 담는다. 빈 배열은 조건이 없다는 뜻이 아니다.",
            "items" : {
              "$ref" : "#/components/schemas/PolicySourceCondition"
            }
          },
          "sourceNotices" : {
            "type" : "array",
            "description" : "현재 수집 내용에서 확인한 조건 충돌 안내. 빈 배열은 검토 완료를 뜻하지 않는다.",
            "items" : {
              "$ref" : "#/components/schemas/PolicySourceNotice"
            }
          },
          "sourceUrl" : {
            "type" : "string"
          }
        },
        "required" : [ "collectedAt", "content", "policyNumber", "recruitment", "revision", "sourceConditions", "sourceNotices", "sourceUrl" ]
      },
      "PolicyEvaluatedCheck" : {
        "type" : "object",
        "properties" : {
          "evidence" : {
            "type" : "string"
          },
          "explanation" : {
            "type" : "string"
          },
          "label" : {
            "type" : "string"
          },
          "outcome" : {
            "type" : "string",
            "enum" : [ "MET", "NOT_MET", "UNKNOWN" ]
          },
          "providedValue" : {
            "type" : "string"
          }
        },
        "required" : [ "evidence", "explanation", "label", "outcome", "providedValue" ]
      },
      "PolicyEvaluation" : {
        "type" : "object",
        "properties" : {
          "checks" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/PolicyEvaluatedCheck"
            }
          },
          "commonCriteriaStatus" : {
            "type" : "string",
            "enum" : [ "ELIGIBLE", "NEEDS_REVIEW", "INELIGIBLE" ]
          },
          "evaluatedAt" : {
            "type" : "string",
            "format" : "date-time"
          },
          "explanation" : {
            "type" : "string"
          },
          "policyNumber" : {
            "type" : "string"
          },
          "remainingChecks" : {
            "type" : "array",
            "items" : {
              "type" : "string"
            }
          },
          "revision" : {
            "type" : "integer",
            "format" : "int64"
          },
          "ruleVersion" : {
            "type" : "string"
          },
          "scope" : {
            "type" : "string"
          },
          "sourceUrl" : {
            "type" : "string"
          },
          "status" : {
            "type" : "string",
            "enum" : [ "ELIGIBLE", "NEEDS_REVIEW", "INELIGIBLE" ]
          }
        },
        "required" : [ "checks", "commonCriteriaStatus", "evaluatedAt", "explanation", "policyNumber", "remainingChecks", "revision", "ruleVersion", "scope", "sourceUrl", "status" ]
      },
      "PolicyEvaluationRequest" : {
        "type" : "object",
        "properties" : {
          "answers" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/PolicyAnswer"
            },
            "maxItems" : 20,
            "minItems" : 0
          },
          "revision" : {
            "type" : "integer",
            "format" : "int64",
            "minimum" : 0
          },
          "ruleVersion" : {
            "type" : "string",
            "maxLength" : 80,
            "minLength" : 0
          }
        },
        "required" : [ "answers", "revision", "ruleVersion" ]
      },
      "PolicyListResponse" : {
        "type" : "object",
        "properties" : {
          "hasNext" : {
            "type" : "boolean"
          },
          "items" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/PolicySummary"
            }
          },
          "page" : {
            "type" : "integer",
            "format" : "int32",
            "description" : "1부터 시작하는 페이지"
          },
          "pageSize" : {
            "type" : "integer",
            "format" : "int32"
          },
          "total" : {
            "type" : "integer",
            "format" : "int64"
          }
        },
        "required" : [ "hasNext", "items", "page", "pageSize", "total" ]
      },
...
      "PolicyPrefill" : {
        "type" : "object",
        "properties" : {
          "answers" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/PolicyAnswer"
            }
          }
        },
        "required" : [ "answers" ]
      },
      "PolicyPrefillRequest" : {
        "type" : "object",
        "properties" : {
          "birthDate" : {
            "type" : "string",
            "format" : "date"
          },
          "revision" : {
            "type" : "integer",
            "format" : "int64",
            "minimum" : 0
          },
          "ruleVersion" : {
            "type" : "string",
            "maxLength" : 80,
            "minLength" : 0
          }
        },
        "required" : [ "birthDate", "revision", "ruleVersion" ]
      },
      "PolicyQuestion" : {
        "type" : "object",
        "properties" : {
          "help" : {
            "type" : "string"
          },
          "id" : {
            "type" : "string",
            "pattern" : "[a-zA-Z][a-zA-Z0-9_]{0,59}"
          },
          "label" : {
            "type" : "string",
            "minLength" : 1
          },
          "options" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/PolicyAnswerOption"
            },
            "minItems" : 1
          }
        },
        "required" : [ "help", "id", "label", "options" ]
      },
      "PolicyQuestionnaire" : {
        "type" : "object",
        "properties" : {
          "available" : {
            "type" : "boolean"
          },
          "policyNumber" : {
            "type" : "string"
          },
          "questions" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/PolicyQuestion"
            }
          },
          "reason" : {
            "type" : "string"
          },
          "revision" : {
            "type" : "integer",
            "format" : "int64"
          },
          "ruleVersion" : {
            "type" : "string"
          },
          "scope" : {
            "type" : "string"
          },
          "sourceUrl" : {
            "type" : "string"
          }
        },
        "required" : [ "available", "policyNumber", "questions", "reason", "revision", "ruleVersion", "scope", "sourceUrl" ]
      },
      "PolicyRecruitment" : {
        "type" : "object",
        "properties" : {
          "daysUntilDeadline" : {
            "type" : [ "integer", "null" ],
            "format" : "int32",
            "description" : "평가 시점의 서울 날짜부터 마감일까지 남은 일수. 마감일이 없으면 null"
          },
          "deadlineOnSeoul" : {
            "type" : [ "string", "null" ],
            "format" : "date",
            "description" : "확인한 마감일(서울). 상시·소진형·기간 미확인과 날짜 없는 마감은 null"
          },
          "evaluatedAt" : {
            "type" : "string",
            "format" : "date-time"
          },
          "explanation" : {
            "type" : "string"
          },
          "status" : {
            "type" : "string",
            "enum" : [ "BEFORE_OPENING", "OPEN", "CLOSED", "ROLLING", "UNTIL_EXHAUSTED", "UNKNOWN" ]
          }
        },
        "required" : [ "daysUntilDeadline", "deadlineOnSeoul", "evaluatedAt", "explanation", "status" ]
      },
...
      "PolicySourceCondition" : {
        "type" : "object",
        "properties" : {
          "label" : {
            "type" : "string",
            "description" : "조건 이름. 연령·소득·취업 상태·학력"
          },
          "value" : {
            "type" : "string",
            "description" : "온통청년에 표기된 값"
          }
        },
        "required" : [ "label", "value" ]
      },
      "PolicySourceNotice" : {
        "type" : "object",
        "properties" : {
          "description" : {
            "type" : "string"
          },
          "sourceLabel" : {
            "type" : "string"
          },
          "sourceUrl" : {
            "type" : "string"
          },
          "title" : {
            "type" : "string"
          }
        },
        "required" : [ "description", "sourceLabel", "sourceUrl", "title" ]
      },
      "PolicySummary" : {
        "type" : "object",
        "properties" : {
          "applicationPeriod" : {
            "type" : "string"
          },
          "category" : {
            "type" : "string"
          },
          "collectedAt" : {
            "type" : "string",
            "format" : "date-time"
          },
          "description" : {
            "type" : "string"
          },
          "organization" : {
            "type" : "string"
          },
          "policyNumber" : {
            "type" : "string"
          },
          "questionnaireAvailable" : {
            "type" : "boolean",
            "description" : "현재 원문·적용 시점에 맞는 공통요건 질문 제공 여부. 자격·모집 상태와 무관"
          },
          "recruitment" : {
            "$ref" : "#/components/schemas/PolicyRecruitment"
          },
          "title" : {
            "type" : "string"
          }
        },
        "required" : [ "applicationPeriod", "category", "collectedAt", "description", "organization", "policyNumber", "questionnaireAvailable", "recruitment", "title" ]
      },
...
    },
    "securitySchemes" : {
      "memberSession" : {
        "description" : "카카오·네이버 로그인 후 발급하는 서버 세션",
        "in" : "cookie",
        "name" : "YPM_SESSION",
        "type" : "apiKey"
      }
    }
  }
}